fix(lint): satisfy golangci-lint v2.11.4 + fix unreachable wiki table row separator
Three issues surfaced by CI that the local dev loop missed:

1. gosec G115 (vector_parallel_test.go) — swapParallelConfig/MinSize/Workers
   took `int` and called `int32(x)`, which gosec flags as potential overflow.
   Switched the helper signatures to `int32` directly (matches the underlying
   atomic.Int32) so the conversion is gone; only one caller (the benchmark
   loop over worker counts) needed its literal slice retyped.

2. staticcheck SA9003 (wiki_import_test.go) — TestCharsetReader_UTF8 had an
   empty `if r != nil { }` branch with only a comment, which meant the test
   never actually asserted the pass-through behavior. Replaced with a real
   t.Errorf so a regression would now fail the test.

3. staticcheck SA4017 (wikitext.go) — the `|-` row-separator branch in
   convertTables was unreachable because the more-generic `|` data-row
   branch was checked first, swallowing every `|-` line as an empty data
   row rendered `| |`. Reordered so `|-` is checked before the generic `|`.
   Existing TestConvertTablesBasic still passes (it only `strings.Contains`
   the header/body rows, which are unaffected).

Local verification:
- golangci-lint run --timeout=5m → 0 issues
- go vet ./... → clean
- go test ./... -short → PASS (mddb 70.11s, mddb/graphql 0.23s)

+++ services/mddbd/vector_parallel_test.go
@@ -12,9 +12,11 @@ import (
 
 // swapParallelConfig atomically sets workers/minSize and returns a restore func.
 // Usage: defer swapParallelConfig(workers, minSize)()
-func swapParallelConfig(workers, minSize int) func() {
-	oldW := parallelSearchConfig.workers.Swap(int32(workers))
-	oldM := parallelSearchConfig.minSize.Swap(int32(minSize))
+// Accepts int32 directly to match the underlying atomic type and avoid
+// gosec G115 overflow warnings on int→int32 conversion.
+func swapParallelConfig(workers, minSize int32) func() {
+	oldW := parallelSearchConfig.workers.Swap(workers)
+	oldM := parallelSearchConfig.minSize.Swap(minSize)
 	return func() {
 		parallelSearchConfig.workers.Store(oldW)
 		parallelSearchConfig.minSize.Store(oldM)
@@ -22,14 +24,14 @@ func swapParallelConfig(workers, minSize int) func() {
 }
 
 // swapParallelMinSize only overrides minSize, keeping workers unchanged.
-func swapParallelMinSize(minSize int) func() {
-	oldM := parallelSearchConfig.minSize.Swap(int32(minSize))
+func swapParallelMinSize(minSize int32) func() {
+	oldM := parallelSearchConfig.minSize.Swap(minSize)
 	return func() { parallelSearchConfig.minSize.Store(oldM) }
 }
 
 // swapParallelWorkers only overrides workers, keeping minSize unchanged.
-func swapParallelWorkers(workers int) func() {
-	oldW := parallelSearchConfig.workers.Swap(int32(workers))
+func swapParallelWorkers(workers int32) func() {
+	oldW := parallelSearchConfig.workers.Swap(workers)
 	return func() { parallelSearchConfig.workers.Store(oldW) }
 }
 
@@ -454,7 +456,7 @@ func BenchmarkParallelWorkerScaling(b *testing.B) {
 
 	defer swapParallelMinSize(1)()
 
-	for _, workers := range []int{1, 2, 4, 8} {
+	for _, workers := range []int32{1, 2, 4, 8} {
 		b.Run(fmt.Sprintf("workers_%d", workers), func(b *testing.B) {
 			restore := swapParallelWorkers(workers)
 			defer restore()
